+++ src/main/java/com/redis/DemoforredisApplication.java
@@ -11,33 +11,8 @@ import java.util.Collections;
 import java.util.List;
 
 @SpringBootApplication
-@RestController
-@RequestMapping("/product")
+
 public class DemoforredisApplication {
-
-	@Autowired
-	private ProductDao dao;
-	@PostMapping
-	public Product savealldata(@RequestBody Product product){
-
-		return dao.save(product);
-	}
-	@GetMapping("/{all}")
-	public List<Product> getall(){
-
-		return dao.findAll();
-	}
-@GetMapping("/{id}")
-	public Product findproduct(@PathVariable int id){
-
-		return dao.findProductById(id);
-	}
-@DeleteMapping("/{id}")
-	public  String deletedid(@PathVariable int id){
-
-		return dao.deleteproduct(id);
-	}
-
 	public static void main(String[] args) {
 
 		SpringApplication.run(DemoforredisApplication.class, args);
